working agent

+++ PlayerHealth.cs
@@ -5,9 +5,9 @@ using UnityEngine;
 public class PlayerHealth : MonoBehaviour
 {
 
-    public int health = 100;
+    public float health = 100;
 
-    public int maxHealth = 100;
+    public float maxHealth = 100;
     public Transform respawnPoint; // Assign this in the inspector
 
     public bool IsDead = false;
